+++ Src/evaluation.h
@@ -7,9 +7,9 @@ class Evaluation
 public:
     static const std::map<PieceName, int> pieceNameToValue;
     static float EvaluateMaterial(const Board &board, PieceColors sideToEvaluate);
-    static int EvaluateHangingPieces(const Board &board, PieceColors sideToEvaluate);
-    static int EvaluateDefendedPieces(const Board &board, PieceColors sideToEvaluate);
-    static int EvaluateKingSafety(const Board &board, PieceColors sideToEvaluate);
-    static float EvaluateSide(const Board &board, PieceColors sideToEvaluate);
-    static float Evaluate(const Board &board);
+    static double EvaluateHangingPieces(const Board &board, PieceColors sideToEvaluate);
+    static double EvaluateDefendedPieces(const Board &board, PieceColors sideToEvaluate);
+    static double EvaluateKingSafety(const Board &board, PieceColors sideToEvaluate);
+    static double EvaluateSide(const Board &board, PieceColors sideToEvaluate);
+    static double Evaluate(const Board &board);
 };
